2018.01.28
* Updated to Clover 4394
* Update FakeSMC to support Sierra and High Sierra
* Updated Utility espmount.bash

+++ Utilities/root/usr/local/bin/espmount.bash
@@ -1,9 +1,10 @@
 #!/bin/bash
-if [[ "$1" == "" ]]; then
-echo enter disk number
-read n 
+if [[ "$1" == "" ]]; 
+then
+    echo enter disk number
+    read n 
 else
-n=$1
+    n=$1
 fi;
 
 while [[ $# > 1 ]]
@@ -22,11 +23,37 @@ esac
 shift # past argument or value
 done
 
+# find esp of boot volume
 bvs=$( system_profiler SPSoftwareDataType | grep Volume )
-bv=${bvs#*:}
-esp=$( diskutil list |grep "HFS${bv} "|awk '{print $6}'| sed -e's/s2//' -e's/disk//' )
+bv=${bvs#*: }
+lds=$( diskutil info $bv  |  grep "Part of"          | sed -e's/^.* disk/disk/' )
+pds=$( diskutil list $lds | egrep " Volume on| Store | Container $lds " | sed -e's/^.* disk/disk/' )
+if [[ $pds == '' ]]; 
+then
+    pd=${lds}
+else
+    pd=${pds%s*}
+fi
+besp=${pd#*disk}
 
-if [[ $n == $esp ]]; then
+# find esp of requested volume
+lds=$( diskutil info disk$n |  grep "Part of"          | sed -e's/^.* disk/disk/' )
+pds=$( diskutil list $lds   | egrep " Volume on| Store | Container $lds " | sed -e's/^.* disk/disk/' )
+if [[ $pds == '' ]]; 
+then
+    pd=${lds}
+else
+    pd=${pds%s*}
+fi
+nesp=${pd#*disk}
+
+# if n is a LV vs PV use pd
+if [[ $n == ${lds/disk/} ]];
+then
+    n=$nesp
+fi
+
+if [[ $n == $besp ]]; then
 	p=/Volumes/ESP
 else
 	p=/Volumes/ESP-$n
@@ -34,27 +61,29 @@ fi
 
 
 d=/dev/disk${n}s1
-if [[ $um == Yes ]]; then
+if [[ $um == Yes ]]; 
+then
 	sudo umount -f $d >& /dev/null
 	sleep 2
-	test -e $p && rmdir $p
+	test -e $p && sudo rmdir $p
 	sleep 1
 else
 	mount|grep $d|grep "umes/E" >& /dev/null
-	if [[ $? != 0 ]]; then
-	test -e $p || sudo mkdir $p
-	sudo mount -t msdos -o noowners,noasync,sync $d $p
-	if [[ $? != 0 ]]
-	then
-	sudo rmdir $p
-	exit 1
+	if [[ $? != 0 ]]; 
+        then
+	    test -e $p || sudo mkdir $p
+	    sudo mount -t msdos -o noowners,noasync,sync $d $p
+	    if [[ $? != 0 ]]
+	    then
+	        sudo rmdir $p
+	        exit 1
+	    else
+	        echo $p
+	    fi
+	    exit 0
 	else
-	echo $p
-	fi
-	exit 0
-	else
-	echo $p
-	exit 1
+	    printf "$p\n"
+	    exit 0
 	fi
 fi
